make timestamp an integer in event entities

+++ src/Events/Error.php
@@ -58,7 +58,7 @@ class Error
 
     /**
      * Timestamp in microseconds.
-     * @var float|null
+     * @var int|null
      */
     protected $timestamp;
 
@@ -191,17 +191,17 @@ class Error
     }
 
     /**
-     * @return float|null
+     * @return int|null
      */
-    public function getTimestamp(): ?float
+    public function getTimestamp(): ?int
     {
         return $this->timestamp;
     }
 
     /**
-     * @param float|null $timestamp
+     * @param int|null $timestamp
      */
-    public function setTimestamp(?float $timestamp): void
+    public function setTimestamp(?int $timestamp): void
     {
         $this->timestamp = $timestamp;
     }
